--- src/main/java/ru/skypro/homework/dto/Login.java
@@ -1,26 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-@Data
-public class Login {
-
-    private String username;
-    private String password;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
