--- src/main/java/ru/skypro/homework/dto/Register.java
@@ -1,14 +1,0 @@
-package ru.skypro.homework.dto;
-
-import lombok.Data;
-
-@Data
-public class Register {
-
-    private String username;
-    private String password;
-    private String firstName;
-    private String lastName;
-    private String phone;
-    private Role role;
-}
