+++ src/main/java/ru/skypro/homework/dto/autoAndReg/Login.java
@@ -6,13 +6,29 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
 @Data
-public class authorizationDTO {
+public class Login {
 
     @JsonProperty("username")
     private String username;
 
     @JsonProperty("password")
     private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
 
 
 }
