+++ src/main/java/ru/skypro/homework/service/AuthService.java
@@ -1,6 +1,6 @@
 package ru.skypro.homework.service;
 
-import ru.skypro.homework.dto.Register;
+import ru.skypro.homework.dto.autoAndReg.Register;/*изначально эта зависимость была на файл в папке dto*/
 
 public interface AuthService {
     boolean login(String userName, String password);
