+++ src/main/java/ru/skypro/homework/service/impl/AuthServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
-import ru.skypro.homework.dto.Register;
+import ru.skypro.homework.dto.autoAndReg.Register;
 import ru.skypro.homework.service.AuthService;
 
 @Service
@@ -31,7 +31,7 @@ public class AuthServiceImpl implements AuthService {
 
     @Override
     public boolean register(Register register) {
-        if (manager.userExists(register.getUsername())) {
+        if (manager.userExists(register.getUsername())) {/*Методов getUsername, getPassword, getRole не ьыло в Registrer. Они добавлены отдельно*/
             return false;
         }
         manager.createUser(
